requirements.txt
# NEXA DEEP INTELLIGENCE v5.0 — requirements.txt

# Async HTTP
httpx>=0.27.0

# Telegram Bot
python-telegram-bot[ext]>=21.0.0

# Zamanlayıcı
APScheduler>=3.10.4
pytz>=2024.1

# Veritabanı (async SQLite)
aiosqlite>=0.20.0

# Ortam değişkenleri
python-dotenv>=1.0.0

# Zaman dilimi desteği
tzdata>=2024.1
